requirements.txt
# SHL Assessment Recommender Dependencies

# Web Framework
fastapi>=0.104.0
uvicorn>=0.23.2
streamlit>=1.28.0

# Data Processing
pandas>=2.1.1
numpy>=1.26.0

# Web Scraping
requests>=2.31.0
beautifulsoup4>=4.12.2

# Visualization
plotly>=5.18.0

# Environment Variables
python-dotenv>=1.0.0

# API Integration
pydantic>=2.4.2